fixing greptile issues and refactoring
